+++ src/zarr.hh
@@ -16,10 +16,12 @@
 #include <utility> // std::pair
 #include <vector>
 
+#include <nlohmann/json.hpp>
+
 namespace fs = std::filesystem;
+using json = nlohmann::json;
 
 namespace acquire::sink::zarr {
-
 struct Zarr : public Storage
 {
   public:
@@ -91,6 +93,7 @@ struct Zarr : public Storage
     virtual void write_group_metadata_() const = 0;
 
     /// Multiscale
+    json make_multiscale_metadata_() const;
     void write_multiscale_frames_(const uint8_t* data_,
                                   size_t bytes_of_data,
                                   const ImageShape& shape_);