+++ tests/get-meta.cpp
@@ -86,10 +86,9 @@ main()
                 CHECK(Device_Ok == storage_get_meta(storage, &metadata));
 
                 CHECK(metadata.chunking_is_supported);
+                CHECK(metadata.multiscale_is_supported);
                 CHECK(metadata.s3_is_supported);
                 CHECK((bool)metadata.sharding_is_supported ==
-                      name.starts_with("ZarrV3"));
-                CHECK((bool)metadata.multiscale_is_supported !=
                       name.starts_with("ZarrV3"));
 
                 CHECK(Device_Ok == driver_close_device(device));